update file format diagram docstring and make version a u16 again

+++ vortex-serde/src/layouts/mod.rs
@@ -5,7 +5,7 @@ mod pruning;
 #[cfg(test)]
 mod tests;
 
-pub const VERSION: u32 = 1;
+pub const VERSION: u16 = 1;
 pub const MAGIC_BYTES: [u8; 4] = *b"VRTX";
 // Size of serialized Postscript Flatbuffer
 pub const FOOTER_POSTSCRIPT_SIZE: usize = 32;
